[grug] Restore shared synthetic dataset contract
